add controller implimentation with Response Entity(http status) and Exception Handling

+++ src/main/java/com/spring/tuto/conferencedemo/controllers/SessionController.java
@@ -1,15 +1,23 @@
 package com.spring.tuto.conferencedemo.controllers;
 
 import com.spring.tuto.conferencedemo.DTOs.SessionDTO;
+import com.spring.tuto.conferencedemo.DTOs.responses.AppResponse;
+import com.spring.tuto.conferencedemo.DTOs.responses.ErrorResponse;
+import com.spring.tuto.conferencedemo.DTOs.responses.SessionDetailsResponse;
+import com.spring.tuto.conferencedemo.DTOs.responses.SuccessResponse;
+import com.spring.tuto.conferencedemo.exceptions.SessionNotFoundException;
 import com.spring.tuto.conferencedemo.mapper.SessionMapper;
 import com.spring.tuto.conferencedemo.models.Session;
 import com.spring.tuto.conferencedemo.repositories.SessionRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.function.EntityResponse;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
@@ -23,14 +31,19 @@ public class SessionController {
 
     @GetMapping
     public List<SessionDTO> list() {
-        List<Session> sessions = sessionRepository.findAll();
+        List<Session> sessions = this.sessionRepository.findAll();
         return sessions.stream().map(x->mapper.toSessionDTO(x)).collect(Collectors.toList());
     }
 
      @GetMapping
      @RequestMapping("{id}")
-    public SessionDTO get(@PathVariable Long id){
-        return mapper.toSessionDTO(sessionRepository.getOne(id));
+    public ResponseEntity<AppResponse> get(@PathVariable Long id){
+        Optional<SessionDTO> resp = Optional.ofNullable(mapper.toSessionDTO(sessionRepository.getOne(id)));
+        if(resp.isPresent()){
+            return new ResponseEntity<>(new SessionDetailsResponse(resp.get()),HttpStatus.OK);
+        }
+            return new ResponseEntity<>( new ErrorResponse("Bad request"),HttpStatus.NOT_FOUND);
+
  }
 
     @PostMapping
@@ -40,16 +53,36 @@ public class SessionController {
 
     }
 
+    //@PutMapping("{id}")
     @RequestMapping(method= RequestMethod.PUT, value="{id}")
-    public SessionDTO update(@RequestBody SessionDTO sessionDto , @PathVariable Long id){
+    public ResponseEntity<AppResponse>  update(@RequestBody SessionDTO sessionDto , @PathVariable Long id){
+           Optional<Session> optionalSession = Optional.of(sessionRepository.getOne(id));
+      //another way to show how to deal with exception, we could return new ErrorResponse and httpstatus as NotFound
+        if(!optionalSession.isPresent())throw new SessionNotFoundException();
+
         Session existingSession = sessionRepository.getOne(id);
         BeanUtils.copyProperties(mapper.toSession(sessionDto),existingSession, "session_id");
        Session sessionUpdated = sessionRepository.saveAndFlush(existingSession);
-       return (mapper.toSessionDTO(sessionUpdated));
+        SessionDTO sessionDTO =mapper.toSessionDTO(sessionUpdated);
+       return ResponseEntity.ok(new SessionDetailsResponse((sessionDTO),"session updated succefully"));
     }
 
-    @RequestMapping(method = RequestMethod.DELETE, value ="{id}")
-    public void delete(@PathVariable Long id){
-        sessionRepository.deleteById(id);
+    @DeleteMapping("{id}")
+    //@RequestMapping(method = RequestMethod.DELETE, value ="{id}")
+    public ResponseEntity<AppResponse> delete(@PathVariable Long id){
+        Optional<Session> optionalSession = Optional.ofNullable(sessionRepository.getOne(id));
+        if(optionalSession.isPresent()) {
+            sessionRepository.deleteById(id);
+            return ResponseEntity.ok(new SuccessResponse("Session Deleted Succefully"));
+        }else {
+            return new ResponseEntity<>(new ErrorResponse("Session Not Found"), HttpStatus.NOT_FOUND);
+        }
+
+    }
+
+    @DeleteMapping
+    public AppResponse deleteAll(){
+        sessionRepository.deleteAll();
+        return new SuccessResponse("Delete All Sessions Succefully");
     }
 }
